Acrs4 is now a proper class (inherits from std::list). Removed hash_t and plen_t (now uint64/32_t). Combined some overloaded helper functions.

+++ trunk/c++/ip4route.cpp
@@ -118,7 +118,7 @@ namespace IP4Route
             return os;
         }
 
-        os << rt.getNetwork().first << "/" << (int) rt.getPlen() << " in "
+        os << rt.getNetwork().first << "/" << rt.getPlen() << " in "
            << rt.getMetric();
 
         return os;
